more typing of database-collections modules, move name confirmation into state
