+++ packages/databases-collections/src/modules/index.ts
@@ -1,0 +1,7 @@
+import collectionsReducer from './collections';
+import databasesReducer from './databases';
+
+export {
+  collectionsReducer,
+  databasesReducer
+};
