--- packages/databases-collections/src/modules/reset.js
@@ -1,14 +1,0 @@
-/**
- * The reset action name.
- */
-export const RESET = 'databases-collections/reset';
-
-/**
- * Reset the state action.
- *
- * @return {Object} The action creator.
- */
-export const reset = () => ({
-  type: RESET
-});
-
