+++ packages/databases-collections/src/stores/drop-database.js
@@ -1,7 +1,7 @@
 import { createStore, applyMiddleware } from 'redux';
 import thunk from 'redux-thunk';
 import { dataServiceConnected } from '../modules/data-service';
-import reducer, { open } from '../modules/drop-database';
+import reducer, { open } from '../modules/drop-database/drop-database';
 
 const store = createStore(reducer, applyMiddleware(thunk));
